Moved the OTel tests to the SequentialIntegration project to simplify project structure.
