+++ projects/Test/SequentialIntegration/SequentialIntegration.csproj
@@ -16,7 +16,7 @@
   </PropertyGroup>
 
   <ItemGroup>
-    <ProjectReference Include="../../RabbitMQ.Client/RabbitMQ.Client.csproj" />
+    <ProjectReference Include="../../RabbitMQ.Client.OpenTelemetry/RabbitMQ.Client.OpenTelemetry.csproj" />
     <ProjectReference Include="../Common/Common.csproj" />
   </ItemGroup>
 
@@ -36,6 +36,7 @@
   <ItemGroup>
     <PackageReference Include="Microsoft.NET.Test.Sdk" Version="17.9.0" />
     <PackageReference Include="Microsoft.NETFramework.ReferenceAssemblies" Version="1.0.3" PrivateAssets="all" />
+    <PackageReference Include="OpenTelemetry.Exporter.InMemory" Version="1.8.0" />
     <PackageReference Include="xunit" Version="2.7.0" />
     <PackageReference Include="xunit.runner.visualstudio" Version="2.5.7" PrivateAssets="all" />
   </ItemGroup>
